+++ src/webapp/src/utils/plugins.js
@@ -8,5 +8,15 @@ export default {
       )
       return isPluginInstalled || !plugin.hidden
     })
+  },
+  filterAvailablePlugins({ installedPlugins = [], pluginList = [] }) {
+    return pluginList.filter(plugin => {
+      const isPluginInstalled = Boolean(
+        installedPlugins.find(
+          installedPlugin => installedPlugin.name === plugin.name
+        )
+      )
+      return !isPluginInstalled && !plugin.hidden
+    })
   }
 }